+++ python/clu/command.py
@@ -14,10 +14,22 @@ import sys
 import time
 from contextlib import suppress
 
-from typing import Any, Callable, Dict, Generic, List, Optional, Tuple, TypeVar, Union
+from typing import (
+    Any,
+    Awaitable,
+    Callable,
+    Dict,
+    Generic,
+    List,
+    Optional,
+    Tuple,
+    TypeVar,
+    Union,
+)
 
 import clu
 import clu.base
+from clu.exceptions import CommandError
 from clu.tools import CommandStatus, StatusMixIn
 
 
@@ -91,7 +103,7 @@ class BaseCommand(
 
     def __init__(
         self,
-        commander_id: Union[int, str] = 0,
+        commander_id: Union[int, str, None] = None,
         command_id: Union[int, str] = 0,
         consumer_id: Union[int, str] = 0,
         actor: Optional[Actor_co] = None,
@@ -298,6 +310,26 @@ class BaseCommand(
             **kwargs,
         )
 
+    def send_command(
+        self,
+        target: str,
+        command_string: str,
+        *args,
+        **kwargs,
+    ) -> BaseCommand | Awaitable[Command]:
+        """Sends a command to an actor using the commander ID of this command."""
+
+        if self.actor is None:
+            raise CommandError("The actor need to be defined to send commands.")
+
+        return self.actor.send_command(
+            target,
+            command_string,
+            *args,
+            command=self,
+            **kwargs,
+        )
+
 
 class Command(BaseCommand[Actor_co, "Command"]):
     """A command from a user.
@@ -352,7 +384,7 @@ class Command(BaseCommand[Actor_co, "Command"]):
         )
 
 
-def parse_legacy_command(command_string: str) -> Tuple[int, str]:
+def parse_legacy_command(command_string: str) -> Tuple[Union[str, None], int, str]:
     """Parses a command received by a legacy actor.
 
     Parameters
@@ -369,24 +401,26 @@ def parse_legacy_command(command_string: str) -> Tuple[int, str]:
     """
 
     _HEADER_BODY_RE = re.compile(
-        r"((?P<cmdID>\d+)(?:\s+\d+)?\s+)?((?P<cmdBody>[A-Za-z_].*|--help))?$"
+        r"(?:([a-z0-9]*\.[a-z0-9_\.]+)\s+)?"
+        r"(?:(\d+)(?:\s+\d+)?\s+)?"
+        r"(?:([a-z_].*|--help))?$",
+        re.IGNORECASE,
     )
 
     command_match = _HEADER_BODY_RE.match(command_string)
     if not command_match:
         raise clu.CommandError(f"Could not parse command {command_string!r}")
 
-    command_dict = command_match.groupdict("")
+    commander, command_id_str, command_body = command_match.groups()
 
-    command_id_str = command_dict["cmdID"]
     if command_id_str:
         command_id = int(command_id_str)
     else:
         command_id = 0
 
-    command_body = command_dict.get("cmdBody", "").strip()
+    command_body = command_body.strip()
 
-    return command_id, command_body
+    return commander, command_id, command_body
 
 
 class TimedCommandList(list):
@@ -514,7 +548,7 @@ class TimedCommand(object):
         await Command(
             self.command_string,
             actor=actor,
-            commander_id=actor.name,
+            commander_id=f".{actor.name}",
             silent=silent,
         ).parse()